feat(BUX-177): refctor code, adjust to review
